Bugfix: Don't use agent by default (#151)
Require users to explicitly pass `use_agent=True` to upload via the agent.

Trying to use the agent by default can cause backwards incompatible behavior 
for users with old versions of the agent installed. In this case, new installations 
of the Python client will raise an error because the agent version is invalid.