Allow model values to be arrays (#163)
Refactor the internals of model property types to allow array values to
be correctly set.

This change moves all the logic of validating data types and casting
property values into the `ModelPropertyType` and `ModelPropertyEnumType`
objects in order to special-case situations that are only applicable to
enumerated types. In addition, internal types that were previously
represented by Python literals (`str`, `int`, etc) are uniformly
converted to `ModelPropertyType` instances, making it easier to reason
about the code.